// src/sanity/sanity.types.ts
export type SanityImageCrop = {
  _type: 'sanity.imageCrop';
  top: number;
  bottom: number;
  left: number;
  right: number;
};

export type SanityImageHotspot = {
  _type: 'sanity.imageHotspot';
  x: number;
  y: number;
  height: number;
  width: number;
};
// ...
export type ARTICLES_QUERY_RESULT = Array<{
  _id: string;
  title: string;
  slug: string;
  excerpt: string;
  publishedAt: string;
  image: {
    asset: {
      _id: string;
      url: string;
      metadata: {
        lqip: string | null;
      } | null;
    } | null;
    alt: string;
    crop: SanityImageCrop | null;
    hotspot: SanityImageHotspot | null;
  };
}>;